launch/requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.3
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
multidict==4.5.2
python-dateutil==2.8.0
python-launch-library==0.6
requests==2.22.0
six==1.12.0
Unidecode==1.0.23
urllib3==1.25.3
Werkzeug==0.15.4
yarl==1.3.0
